cleanup encrypter abstraction
