--- service/controller/cluster_resource_set_config.go
@@ -1,63 +1,0 @@
-package controller
-
-import (
-	"net"
-
-	"github.com/giantswarm/certs"
-	"github.com/giantswarm/k8sclient"
-	"github.com/giantswarm/micrologger"
-	"github.com/giantswarm/randomkeys"
-
-	"github.com/giantswarm/aws-operator/client/aws"
-	"github.com/giantswarm/aws-operator/service/controller/resource/tccp"
-	"github.com/giantswarm/aws-operator/service/internal/locker"
-)
-
-type clusterResourceSetConfig struct {
-	CertsSearcher      certs.Interface
-	HostAWSConfig      aws.Config
-	K8sClient          k8sclient.Interface
-	Locker             locker.Interface
-	Logger             micrologger.Logger
-	RandomKeysSearcher randomkeys.Interface
-
-	AccessLogsExpiration       int
-	AdvancedMonitoringEC2      bool
-	APIWhitelist               tccp.ConfigAPIWhitelist
-	CalicoCIDR                 int
-	CalicoMTU                  int
-	CalicoSubnet               string
-	ClusterIPRange             string
-	DockerDaemonCIDR           string
-	GuestAvailabilityZones     []string
-	GuestPrivateSubnetMaskBits int
-	GuestPublicSubnetMaskBits  int
-	GuestSubnetMaskBits        int
-	IncludeTags                bool
-	IgnitionPath               string
-	ImagePullProgressDeadline  string
-	InstallationName           string
-	IPAMNetworkRange           net.IPNet
-	DeleteLoggingBucket        bool
-	ClusterDomain              string
-	NetworkSetupDockerImage    string
-	Route53Enabled             bool
-	RouteTables                string
-	PodInfraContainerImage     string
-	RegistryDomain             string
-	SSHUserList                string
-	SSOPublicKey               string
-	VaultAddress               string
-}
-
-func (c clusterResourceSetConfig) GetInstallationName() string {
-	return c.InstallationName
-}
-
-func (c clusterResourceSetConfig) GetLogger() micrologger.Logger {
-	return c.Logger
-}
-
-func (c clusterResourceSetConfig) GetVaultAddress() string {
-	return c.VaultAddress
-}
