--- service/controller/machine_deployment_resource_set_config.go
@@ -1,59 +1,0 @@
-package controller
-
-import (
-	"net"
-
-	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
-	"github.com/giantswarm/certs"
-	"github.com/giantswarm/micrologger"
-	"github.com/giantswarm/randomkeys"
-	"k8s.io/client-go/kubernetes"
-
-	"github.com/giantswarm/aws-operator/client/aws"
-	"github.com/giantswarm/aws-operator/service/internal/locker"
-)
-
-type machineDeploymentResourceSetConfig struct {
-	CertsSearcher      certs.Interface
-	G8sClient          versioned.Interface
-	K8sClient          kubernetes.Interface
-	Locker             locker.Interface
-	Logger             micrologger.Logger
-	RandomKeysSearcher randomkeys.Interface
-
-	CalicoCIDR                 int
-	CalicoMTU                  int
-	CalicoSubnet               string
-	ClusterIPRange             string
-	DockerDaemonCIDR           string
-	GuestPrivateSubnetMaskBits int
-	GuestPublicSubnetMaskBits  int
-	GuestSubnetMaskBits        int
-	HostAWSConfig              aws.Config
-	IgnitionPath               string
-	ImagePullProgressDeadline  string
-	InstallationName           string
-	IPAMNetworkRange           net.IPNet
-	ClusterDomain              string
-	NetworkSetupDockerImage    string
-	PodInfraContainerImage     string
-	ProjectName                string
-	RegistryDomain             string
-	Route53Enabled             bool
-	RouteTables                string
-	SSHUserList                string
-	SSOPublicKey               string
-	VaultAddress               string
-}
-
-func (c machineDeploymentResourceSetConfig) GetInstallationName() string {
-	return c.InstallationName
-}
-
-func (c machineDeploymentResourceSetConfig) GetLogger() micrologger.Logger {
-	return c.Logger
-}
-
-func (c machineDeploymentResourceSetConfig) GetVaultAddress() string {
-	return c.VaultAddress
-}
