--- service/controller/spec.go
@@ -1,9 +1,0 @@
-package controller
-
-import "github.com/giantswarm/micrologger"
-
-type encrypterConfigGetter interface {
-	GetInstallationName() string
-	GetLogger() micrologger.Logger
-	GetVaultAddress() string
-}
